refactor: KEEP-1587 move db/schema-extensions.ts to lib/db/
Schema extensions belong with the rest of the DB schema in lib/db/,
not in a standalone db/ directory at the root. Update all imports
from @/db/schema-extensions to @/lib/db/schema-extensions, and use
a relative import in lib/db/schema.ts.
