+++ lib/db/schema-extensions.ts
