+++ lib/web3/nonce-manager.ts
@@ -21,7 +21,7 @@
 import { and, eq, sql } from "drizzle-orm";
 import type { ethers } from "ethers";
 import postgres from "postgres";
-import { pendingTransactions, walletLocks } from "@/db/schema-extensions";
+import { pendingTransactions, walletLocks } from "@/lib/db/schema-extensions";
 import { db } from "@/lib/db";
 import { getDatabaseUrl } from "@/lib/db/connection-utils";
