+++ scripts/miscellaneous/fix-usds-logo.ts
@@ -5,7 +5,7 @@ import "dotenv/config";
 import { eq } from "drizzle-orm";
 import { drizzle } from "drizzle-orm/postgres-js";
 import postgres from "postgres";
-import { supportedTokens } from "@/db/schema-extensions";
+import { supportedTokens } from "@/lib/db/schema-extensions";
 
 const CORRECT_USDS_LOGO =
   "https://raw.githubusercontent.com/trustwallet/assets/master/blockchains/ethereum/assets/0xdC035D45d973E3EC169d2276DDab16f1e407384F/logo.png";
