+++ scripts/seed/seed-tokens.ts
@@ -18,7 +18,7 @@ import { and, eq } from "drizzle-orm";
 import { drizzle } from "drizzle-orm/postgres-js";
 import { ethers } from "ethers";
 import postgres from "postgres";
-import { supportedTokens } from "@/db/schema-extensions";
+import { supportedTokens } from "@/lib/db/schema-extensions";
 import ERC20_ABI from "../../lib/contracts/abis/erc20.json";
 import { getDatabaseUrl } from "../../lib/db/connection-utils";
 import { getRpcUrlByChainId } from "../../lib/rpc/rpc-config";
